change(server) model posture joints type to int

+++ TechnicalResearch/server/app/models/postures_model.py
@@ -4,11 +4,11 @@ from pydantic import BaseModel
 class Postures(BaseModel):
     pos_id: str
     pos_name: str
-    j1: float
-    j2: float
-    j3: float
-    j4: float
-    j5: float
-    j6: float
+    j1: int
+    j2: int
+    j3: int
+    j4: int
+    j5: int
+    j6: int
     angle: int
     gap: int
